+++ marketplace/management/commands/find_duplicates.py
@@ -1,5 +1,5 @@
 from django.core.management.base import BaseCommand
-from django.contrib.auth.models import User
+from users.model import CustomUser
 from marketplace.models import WishList
 
 
@@ -7,7 +7,7 @@ class Command(BaseCommand):
     help = "Find dublicate wishlist objects"
 
     def handle(self, *args, **kwargs):
-        users = User.objects.all()
+        users = CustomUser.objects.all()
         for user in users:
             wishlist = WishList.objects.filter(user=user)